Add test_bases from QuantumOpticsBase and reorganize a bit
