+++ src/linalg.jl
@@ -1,10 +1,194 @@
-samebases(a::AbstractOperator) = samebases(a.basis_l, a.basis_r)::Bool
-samebases(a::AbstractOperator, b::AbstractOperator) = samebases(a.basis_l, b.basis_l)::Bool && samebases(a.basis_r, b.basis_r)::Bool
-check_samebases(a::Union{AbstractOperator, AbstractSuperOperator}) = check_samebases(a.basis_l, a.basis_r)
-multiplicable(a::AbstractOperator, b::AbstractOperator) = multiplicable(a.basis_r, b.basis_l)
-dagger(a::AbstractOperator) = arithmetic_unary_error("Hermitian conjugate", a)
-transpose(a::AbstractOperator) = arithmetic_unary_error("Transpose", a)
-directsum(a::AbstractOperator...) = reduce(directsum, a)
+##
+# Basis checks
+##
+
+const BASES_CHECK = Ref(true)
+
+"""
+    @samebases
+
+Macro to skip checks for same bases. Useful for `*`, `expect` and similar
+functions.
+"""
+macro samebases(ex)
+    return quote
+        BASES_CHECK.x = false
+        local val = $(esc(ex))
+        BASES_CHECK.x = true
+        val
+    end
+end
+
+"""
+    samebases(a, b)
+
+Test if two objects have the same bases.
+"""
+samebases(b1::Basis, b2::Basis) = b1==b2
+samebases(b1::Tuple{Basis, Basis}, b2::Tuple{Basis, Basis}) = b1==b2 # for checking superoperators
+
+"""
+    check_samebases(a, b)
+
+Throw an [`IncompatibleBases`](@ref) error if the objects don't have
+the same bases.
+"""
+function check_samebases(b1, b2)
+    if BASES_CHECK[] && !samebases(b1, b2)
+        throw(IncompatibleBases())
+    end
+end
+
+
+"""
+    multiplicable(a, b)
+
+Check if two objects are multiplicable.
+"""
+multiplicable(b1::Basis, b2::Basis) = b1==b2
+
+function multiplicable(b1::CompositeBasis, b2::CompositeBasis)
+    if !equal_shape(b1.shape,b2.shape)
+        return false
+    end
+    for i=1:length(b1.shape)
+        if !multiplicable(b1.bases[i], b2.bases[i])
+            return false
+        end
+    end
+    return true
+end
+
+"""
+    check_multiplicable(a, b)
+
+Throw an [`IncompatibleBases`](@ref) error if the objects are
+not multiplicable.
+"""
+function check_multiplicable(b1, b2)
+    if BASES_CHECK[] && !multiplicable(b1, b2)
+        throw(IncompatibleBases())
+    end
+end
+
+samebases(a::AbstractOperator) = samebases(a.basis_l, a.basis_r)::Bool # FIXME issue #12
+samebases(a::AbstractOperator, b::AbstractOperator) = samebases(a.basis_l, b.basis_l)::Bool && samebases(a.basis_r, b.basis_r)::Bool # FIXME issue #12
+check_samebases(a::Union{AbstractOperator, AbstractSuperOperator}) = check_samebases(a.basis_l, a.basis_r) # FIXME issue #12
+multiplicable(a::AbstractOperator, b::AbstractOperator) = multiplicable(a.basis_r, b.basis_l) # FIXME issue #12
+
+##
+# tensor, reduce, ptrace
+##
+
+tensor(b::Basis) = b
+
+"""
+    tensor(x::Basis, y::Basis, z::Basis...)
+
+Create a [`CompositeBasis`](@ref) from the given bases.
+
+Any given CompositeBasis is expanded so that the resulting CompositeBasis never
+contains another CompositeBasis.
+"""
+tensor(b1::Basis, b2::Basis) = CompositeBasis([length(b1); length(b2)], (b1, b2))
+tensor(b1::CompositeBasis, b2::CompositeBasis) = CompositeBasis([b1.shape; b2.shape], (b1.bases..., b2.bases...))
+function tensor(b1::CompositeBasis, b2::Basis)
+    N = length(b1.bases)
+    shape = vcat(b1.shape, length(b2))
+    bases = (b1.bases..., b2)
+    CompositeBasis(shape, bases)
+end
+function tensor(b1::Basis, b2::CompositeBasis)
+    N = length(b2.bases)
+    shape = vcat(length(b1), b2.shape)
+    bases = (b1, b2.bases...)
+    CompositeBasis(shape, bases)
+end
+tensor(bases::Basis...) = reduce(tensor, bases)
+
+function Base.:^(b::Basis, N::Integer)
+    if N < 1
+        throw(ArgumentError("Power of a basis is only defined for positive integers."))
+    end
+    tensor([b for i=1:N]...)
+end
+
+"""
+    reduced(a, indices)
+
+Reduced basis, state or operator on the specified subsystems.
+
+The `indices` argument, which can be a single integer or a vector of integers,
+specifies which subsystems are kept. At least one index must be specified.
+"""
+function reduced(b::CompositeBasis, indices)
+    if length(indices)==0
+        throw(ArgumentError("At least one subsystem must be specified in reduced."))
+    elseif length(indices)==1
+        return b.bases[indices[1]]
+    else
+        return CompositeBasis(b.shape[indices], b.bases[indices])
+    end
+end
+
+"""
+    ptrace(a, indices)
+
+Partial trace of the given basis, state or operator.
+
+The `indices` argument, which can be a single integer or a vector of integers,
+specifies which subsystems are traced out. The number of indices has to be
+smaller than the number of subsystems, i.e. it is not allowed to perform a
+full trace.
+"""
+function ptrace(b::CompositeBasis, indices)
+    J = [i for i in 1:length(b.bases) if i ∉ indices]
+    length(J) > 0 || throw(ArgumentError("Tracing over all indices is not allowed in ptrace."))
+    reduced(b, J)
+end
+
 ptrace(a::AbstractOperator, index) = arithmetic_unary_error("Partial trace", a)
 _index_complement(b::CompositeBasis, indices) = complement(length(b.bases), indices)
 reduced(a, indices) = ptrace(a, _index_complement(basis(a), indices))
+traceout!(s::StateVector, i) = ptrace(s,i)
+
+##
+# nsubsystems
+##
+
+nsubsystems(s::AbstractKet) = nsubsystems(basis(s))
+nsubsystems(s::AbstractOperator) = nsubsystems(basis(s))
+nsubsystems(b::CompositeBasis) = length(b.bases)
+nsubsystems(b::Basis) = 1
+nsubsystems(::Nothing) = 1 # TODO Exists because of QuantumSavory; Consider removing this and reworking the functions that depend on it. E.g., a reason to have it when performing a project_traceout measurement on a state that contains only one subsystem
+
+##
+# directsum
+##
+
+"""
+    directsum(b1::Basis, b2::Basis)
+
+Construct the [`SumBasis`](@ref) out of two sub-bases.
+"""
+directsum(b1::Basis, b2::Basis) = SumBasis(Int[length(b1); length(b2)], Basis[b1, b2])
+directsum(b::Basis) = b
+directsum(b::Basis...) = reduce(directsum, b)
+function directsum(b1::SumBasis, b2::Basis)
+    shape = [b1.shape;length(b2)]
+    bases = [b1.bases...;b2]
+    return SumBasis(shape, (bases...,))
+end
+function directsum(b1::Basis, b2::SumBasis)
+    shape = [length(b1);b2.shape]
+    bases = [b1;b2.bases...]
+    return SumBasis(shape, (bases...,))
+end
+function directsum(b1::SumBasis, b2::SumBasis)
+    shape = [b1.shape;b2.shape]
+    bases = [b1.bases...;b2.bases...]
+    return SumBasis(shape, (bases...,))
+end
+
+directsum(x::StateVector...) = reduce(directsum, x)
+directsum(a::AbstractOperator...) = reduce(directsum, a)
